% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convert_psytools_folder}
\alias{convert_psytools_folder}
\title{convert_psytools_folder}
\usage{
convert_psytools_folder(
  folder_location,
  output_location = folder_location,
  output_suffix = "_proc",
  output_format = "xlsx",
  regexp_filter_ext = "\\\\.csv\\\\.gz$",
  regexp_filter_name = ".*"
)
}
\arguments{
\item{folder_location}{"/path/to/folder/to/scan/"}

\item{output_location}{Where to place output files
(default: same directory as input).}

\item{output_suffix}{Each output file will be named as the input, but
with \code{output_suffix} added.
E.g. the default (\code{"_proc"}) means that
\code{"input/data.csv.gz"} will become
\code{"output/data_proc.csv"}.}

\item{output_format}{\code{"xlsx"}, \code{"csv"},
both (i.e. \code{c("xlsx", "csv")}),
or "none" to suppress output and simply return processed data.}

\item{regexp_filter_ext}{regular expression filter for file extensions.
The default selects all gzipped csv files.
For unzipped csv use \code{"\\.csv$"}.}

\item{regexp_filter_name}{regular expression filter for file names
(is pasted with \code{regexp_filter_ext})}
}
\value{
A list of \code{\link[tibble:tbl_df]{tibble}} objects,
    one per input log file. Objects have been rotated from the input log
    format, such that \code{Trial}s are columns and each row is a
    single acquisition session.
}
\description{
Read all psytools logfiles from a folder (\code{folder_location}),
    rotate/format data (using \code{\link{process_psytools_logs}}) and
    write out results in csv format to a folder specified by
    \code{output_location} (default: input folder).
}
