% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPsytools-package.R
\docType{package}
\name{extractPsytools-package}
\alias{extractPsytools}
\alias{extractPsytools-package}
\title{extractPsytools: Functions to Parse Psytools Log Files}
\description{
Reads log files into tidy tabular data, one row per questionnaire.
}
\keyword{internal}
