% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{process_psytools_logs}
\alias{process_psytools_logs}
\title{process_psytools_logs}
\usage{
process_psytools_logs(x)
}
\arguments{
\item{x}{a list of imported psytools logs
(see: \code{\link{read_psytools_logs}}).}
}
\description{
Reshape log file into tabular data. Reading table column names from the
    \code{Trial} column in the log, and associated values from the
    last logged \code{Trial Result} under that \code{Trial}.
}
