% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_psytools_logs}
\alias{read_psytools_logs}
\title{read_psytools_logs}
\usage{
read_psytools_logs(filenames)
}
\arguments{
\item{filenames}{a list of filepaths (e.g. output from
\code{\link{list_psytools_files}})}
}
\value{
A list of \code{\link[tibble:tbl_df]{tibble}} objects,
    one per input file. Convert these from log format to tables with
    \code{\link{process_psytools_logs}}.
}
\description{
Reads in data from a list of Psytools log file locations to a list.
}
\details{
Errors indicate the format is not as expected for a psytools log. Check all
files in the input list are actually Psytools logs.
}
