% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{skipback_bugreport_folder}
\alias{skipback_bugreport_folder}
\title{skipback_bugreport_folder}
\usage{
skipback_bugreport_folder(
  folder_location,
  output_location = folder_location,
  output_suffix = "_skipbackbugreport",
  output_format = "xlsx",
  regexp_filter_ext = "\\\\.csv\\\\.gz$",
  regexp_filter_name = ".*",
  remove_ids = c("EBTEST")
)
}
\arguments{
\item{folder_location}{"/path/to/folder/to/scan/"}

\item{output_location}{Where to place output files
(default: same directory as input).}

\item{output_suffix}{Each output file will be named as the input, but
with \code{output_suffix} added.
E.g. the default (\code{"_skipbackbugreport"}) means that
\code{"input/data.csv.gz"} will become
\code{"output/data_skipbackbugreport.csv"}.}

\item{output_format}{\code{"xlsx"}, \code{"csv"},
both (i.e. \code{c("xlsx", "csv")}),
or "none" to suppress output and simply return processed data.}

\item{regexp_filter_ext}{regular expression filter for file extensions.
The default selects all gzipped csv files.
For unzipped csv use \code{"\\.csv$"}.}

\item{regexp_filter_name}{regular expression filter for file names
(is pasted with \code{regexp_filter_ext})}

\item{remove_ids}{a character vector of IDs to remove.}
}
\description{
Read all psytools logfiles from a folder (\code{folder_location}),
    for each write out an excel-file with data values that would have
    been lost if processed with a previous (bugged) version of the
    psytools processing scripts.
}
\details{
This function was developed to help assess the impact of a retrospective
    script problem identified in 2024. The problem pre-dates the script's
    move to github/r-package.
}
