% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_psytools_logs}
\alias{read_psytools_logs}
\title{read_psytools_logs}
\usage{
read_psytools_logs(filenames, format = c("csv", "txt"))
}
\arguments{
\item{filenames}{a named list of filepaths (e.g. output from
\code{\link{list_psytools_files}})}

\item{format}{how to try to read the data? csv (default) = tabular,
txt = raw strings.}
}
\description{
Reads in psytools log files to a list. Note: errors indicate the csv
    format is not as expected. Try to check the raw files.
}
